\name{ROI.plugin.qpoases-package}
\docType{package}
\alias{ROI.plugin.qpoases-package}

\title{qpOASES}

\description{
This package provides an interface to \code{qpOASES}. 
\code{pOASES} is an open-source \code{C++} implementation of 
the online active set strategy, which was inspired by important observations 
from the field of parametric quadratic programming (QP).
}

\references{
H.J. Ferreau and C. Kirches and A. Potschka and H.G. Bock and M. Diehl.
\code{qpOASES}: A parametric active-set algorithm for quadratic programming,
``Mathematical Programming Computation'', 2014

H.J. Ferreau and H.G. Bock and M. Diehl. 
An online active set strategy to overcome the limitations of explicit MPC,
``International Journal of Robust and Nonlinear Control'', 2008

H.J. Ferreau and A. Potschka and C. Kirches. \code{qpOASES} ``webpage''
\url{https://projects.coin-or.org/qpOASES}, 2017
}
